requirements.txt
onnx==1.16.1
onnxruntime==1.18.0
numpy==1.26.0
colorama==0.4.6
pillow==10.3.0
matplotlib==3.9.0